refactor: move filter.lua out of src/
