+++ lua/neotest-kotlin/init.lua
@@ -2,7 +2,7 @@ local lib = require("neotest.lib")
 local async = require("neotest.async")
 local treesitter = require("neotest-kotlin.treesitter")
 local command = require("neotest-kotlin.src.command")
-local filter = require("neotest-kotlin.src.filter")
+local filter = require("neotest-kotlin.filter")
 local output_parser = require("neotest-kotlin.output_parser")
 
 local M = {}
